+++ config/autoload/global.php
@@ -20,8 +20,4 @@ return [
     'php-debug-bar' => [
         'enabled' => APP_ENV_DEV,
     ],
-    // Bupy7\Mailgun
-    'mailgun' => [
-        'debug' => APP_DEBUG,
-    ],
 ];