+++ module/ExAssetic/config/module.config.php
@@ -9,15 +9,14 @@ namespace ExAssetic;
 return [
     'ex_assetic' => [],
     'service_manager' => [
-        'invokables' => [
-            'AsseticCacheBuster' => \AsseticBundle\CacheBuster\LastModifiedStrategy::class,
-        ],
         'factories' => [
             \Assetic\Filter\Sass\SassFilter::class => Filter\Sass\SassFilterFactory::class,
             \Assetic\Filter\Yui\CssCompressorFilter::class => Filter\Yui\CssCompressorFilterFactory::class,
             \Assetic\Filter\UglifyJs2Filter::class => Filter\UglifyJs2FilterFactory::class,
 
             Options\ModuleOptions::class => Options\ModuleOptionsFactory::class,
+
+            'AsseticCacheBuster' => Cache\CacheBusterFactory::class,
         ],
         'aliases' => [
             'CssCompressorFilter' => \Assetic\Filter\Yui\CssCompressorFilter::class,